general code cleanup in the build system

+++ include/git.sh
@@ -1,34 +1,30 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
-# SPDX-FileCopyrightText: 2020,2021,2023,2024 Leah Rowe <leah@libreboot.org>
-# SPDX-FileCopyrightText: 2022 Caleb La Grange <thonkpeasant@protonmail.com>
+# Copyright (c) 2020-2021,2023-2024 Leah Rowe <leah@libreboot.org>
+# Copyright (c) 2022 Caleb La Grange <thonkpeasant@protonmail.com>
 
 eval "$(setvars "" _target rev _xm loc url bkup_url depend tree_depend xtree \
     mdir subrev subrepo subrepo_bkup)"
 
 fetch_project_trees()
 {
-	_target="${target}"
-	[ ! -d "src/${project}/${project}" ] && x_ mkdir -p "src/${project}" \
-	    && fetch_project_repo "${project}"
+	_target="$target"
+	[ ! -d "src/$project/$project" ] && x_ mkdir -p "src/$project" \
+	    && fetch_project_repo "$project"
 	fetch_config
-	if [ -d "src/${project}/${tree}" ]; then
-		printf "download/%s %s (%s): exists\n" \
-		    "${project}" "${tree}" "${_target}" 1>&2
-		return 0
-	fi
+	e "src/$project/$tree" d && return 0
 	prepare_new_tree
 }
 
 fetch_config()
 {
-	rm -f "${cfgsdir}/"*/seen || $err "fetch_config ${cfgsdir}: !rm seen"
+	rm -f "$cfgsdir/"*/seen || $err "fetch_config $cfgsdir: !rm seen"
 	eval "$(setvars "" xtree tree_depend)"
 	while true; do
 		eval "$(setvars "" rev tree)"
-		_xm="fetch_config ${project}/${_target}"
-		load_target_config "${_target}"
-		[ "${_target}" = "${tree}" ] && break
-		_target="${tree}"
+		_xm="fetch_config $project/$_target"
+		load_target_config "$_target"
+		[ "$_target" = "$tree" ] && break
+		_target="$tree"
 	done
 	[ -n "$tree_depend" ] && [ "$tree_depend" != "$tree" ] && \
 		x_ ./update trees -f "$project" "$tree_depend"; return 0
@@ -37,8 +33,7 @@ fetch_config()
 load_target_config()
 {
 	[ -f "$cfgsdir/$1/target.cfg" ] || $err "$1: target.cfg missing"
-	[ -f "${cfgsdir}/${1}/seen" ] && \
-		$err "${_xm} check: infinite loop in tree definitions"
+	[ -f "$cfgsdir/$1/seen" ] && $err "$_xm cfg: infinite loop in trees"
 
 	. "$cfgsdir/$1/target.cfg" || $err "load_target_config !$cfgsdir/$1"
 	touch "$cfgsdir/$1/seen" || $err "load_config $cfgsdir/$1: !mk seen"
@@ -48,8 +43,8 @@ prepare_new_tree()
 {
 	printf "Creating %s tree %s (%s)\n" "$project" "$tree" "$_target"
 
-	cp -R "src/${project}/${project}" "${tmpgit}" || \
-	    $err "prepare_new_tree ${project}/${tree}: can't make tmpclone"
+	cp -R "src/$project/$project" "$tmpgit" || \
+	    $err "prepare_new_tree $project/$tree: can't make tmpclone"
 	git_prep "$PWD/$cfgsdir/$tree/patches" "src/$project/$tree" "update"
 }
 
@@ -57,28 +52,25 @@ fetch_project_repo()
 {
 	eval "$(setvars "" xtree tree_depend)"
 
-	scan_config "${project}" "config/git"
+	scan_config "$project" "config/git"
 	[ -z "${loc+x}" ] && $err "fetch_project_repo $project: loc not set"
 	[ -z "${url+x}" ] && $err "fetch_project_repo $project: url not set"
 
 	clone_project
-	[ -z "${depend}" ] || for d in ${depend} ; do
-		x_ ./update trees -f ${d}
+	[ -z "$depend" ] || for d in $depend ; do
+		x_ ./update trees -f $d
 	done
-	rm -Rf "${tmpgit}" || $err "fetch_repo: !rm -Rf ${tmpgit}"
+	rm -Rf "$tmpgit" || $err "fetch_repo: !rm -Rf $tmpgit"
 }
 
 clone_project()
 {
 	loc="${loc#src/}"
-	loc="src/${loc}"
-	if [ -d "${loc}" ]; then
-		printf "%s already exists, so skipping download\n" "$loc" 1>&2
-		return 0
-	fi
+	loc="src/$loc"
+	e "$loc" d && return 0
 
 	git clone $url "$tmpgit" || git clone $bkup_url "$tmpgit" \
-	    || $err "clone_project: could not download ${project}"
+	    || $err "clone_project: could not download $project"
 	git_prep "$PWD/config/$project/patches" "$loc"
 }
 
@@ -108,7 +100,7 @@ prep_submodules()
 {
 	[ -f "$tmpgit/.gitmodules" ] || return 0
 
-	mdir="${PWD}/config/submodule/$project"
+	mdir="$PWD/config/submodule/$project"
 	[ -n "$tree" ] && mdir="$mdir/$tree"
 
 	if [ -f "$mdir/module.list" ]; then
@@ -137,8 +129,7 @@ fetch_submodule()
 	    $err "! . $mcfgdir/module.cfg"
 
 	if [ -n "$subrepo" ] || [ -n "$subrepo_bkup" ]; then
-		[ -n "$subrev" ] || \
-		    $err "$1 as per $mdir: subrev not defined"
+		[ -n "$subrev" ] || $err "$1, $mdir: subrev not defined"
 
 		rm -Rf "$tmpgit/$1" || $err "!rm '$mdir' '$1'"
 		for mod in "$subrepo" "$subrepo_bkup"; do
@@ -149,16 +140,15 @@ fetch_submodule()
 		done
 		[ -d "$tmpgit/$1" ] || $err "!clone $mod $project $mcfgdir $1"
 	else
-		git -C "$tmpgit" submodule update --init --checkout -- "$1" || \
-		    $err "$mdir: !update $1"
+		git -C "$tmpgit" submodule update --init --checkout -- "$1" \
+		    || $err "$mdir: !update $1"
 	fi
 }
 
 patch_submodule()
 {
-	[ -z "$subrev" ] || \
-		git -C "$tmpgit/$1" reset --hard "$subrev" || \
-		    $err "$mdir $1: cannot reset git revision"
+	[ -z "$subrev" ] || git -C "$tmpgit/$1" reset --hard "$subrev" || \
+	    $err "$mdir $1: cannot reset git revision"
 
 	git_am_patches "$tmpgit/$1" "$mdir/${1##*/}/patches"
 }
@@ -167,7 +157,7 @@ git_am_patches()
 {
 	for _patch in "$2/"*; do
 		[ -L "$_patch" ] || [ ! -f "$_patch" ] || git -C "$1" am \
-		    "$_patch" || $err "git_am $1 $2: !git am $_patch"; continue
+		    "$_patch" || $err "$1 $2: !git am $_patch"; continue
 	done
 	for _patches in "$2/"*; do
 		[ ! -L "$_patches" ] && [ -d "$_patches" ] && \
